fix ut error
