+++ pulsar-broker/src/test/java/org/apache/pulsar/client/api/SimpleProducerConsumerTest.java
@@ -37,7 +37,6 @@ import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.node.JsonNodeType;
 import com.google.common.collect.Lists;
 import com.google.common.collect.Sets;
-import com.google.common.util.concurrent.Uninterruptibles;
 import io.netty.buffer.ByteBuf;
 import io.netty.buffer.Unpooled;
 import io.netty.channel.ChannelDuplexHandler;
@@ -49,7 +48,6 @@ import java.nio.ByteBuffer;
 import java.nio.file.Files;
 import java.nio.file.Paths;
 import java.time.Clock;
-import java.time.Duration;
 import java.time.Instant;
 import java.time.ZoneId;
 import java.util.ArrayList;
@@ -110,7 +108,6 @@ import org.apache.pulsar.client.impl.TopicMessageImpl;
 import org.apache.pulsar.client.impl.TypedMessageBuilderImpl;
 import org.apache.pulsar.client.impl.crypto.MessageCryptoBc;
 import org.apache.pulsar.client.impl.schema.writer.AvroWriter;
-import org.apache.pulsar.client.util.ExecutorProvider;
 import org.apache.pulsar.common.allocator.PulsarByteBufAllocator;
 import org.apache.pulsar.common.api.EncryptionContext;
 import org.apache.pulsar.common.api.EncryptionContext.EncryptionKey;
@@ -4880,129 +4877,5 @@ public class SimpleProducerConsumerTest extends ProducerConsumerBase {
         } else {
             return 0;
         }
-    }
-
-    @Test
-    public void testConsumerMessageListenerExecutorIsolation() throws Exception {
-        log.info("-- Starting {} test --", methodName);
-
-        @Cleanup("shutdownNow")
-        ExecutorService executor = Executors.newCachedThreadPool();
-        List<CompletableFuture<Long>> maxConsumeDelayWithDisableIsolationFutures = new ArrayList<>();
-        int loops = 10;
-        long sleepTimeMs = 1000;
-        for (int i = 0; i < loops; i++) {
-            // The first consumer will consume messages with 1s delay,
-            // and the others will consume messages without delay.
-            // The maxConsumeDelayWithDisableIsolation of all consumers
-            // should be greater than sleepTimeMs cause by disable MessageListenerExecutor.
-            CompletableFuture<Long> maxConsumeDelayFuture = startConsumeAndComputeMaxConsumeDelay(
-                    "persistent://my-property/my-ns/testConsumerMessageListenerDisableIsolation-" + i,
-                    "my-sub-testConsumerMessageListenerDisableIsolation-" + i,
-                    i==0 ? Duration.ofMillis(sleepTimeMs): Duration.ofMillis(0),
-                    false,
-                    executor);
-            maxConsumeDelayWithDisableIsolationFutures.add(maxConsumeDelayFuture);
-        }
-
-        List<CompletableFuture<Long>> maxConsumeDelayWhitEnableIsolationFutures = new ArrayList<>();
-        for (int i = 0; i < loops; i++) {
-            // The first consumer will consume messages with 1s delay,
-            // and the others will consume messages without delay.
-            // The maxConsumeDelayWhitEnableIsolation of the first consumer
-            // should be greater than sleepTimeMs, and the others should be
-            // less than sleepTimeMs, cause by enable MessageListenerExecutor.
-            CompletableFuture<Long> maxConsumeDelayFuture = startConsumeAndComputeMaxConsumeDelay(
-                    "persistent://my-property/my-ns/testConsumerMessageListenerEnableIsolation-" + i,
-                    "my-sub-testConsumerMessageListenerEnableIsolation-" + i,
-                    i==0 ? Duration.ofMillis(sleepTimeMs): Duration.ofMillis(0),
-                    true,
-                    executor);
-            maxConsumeDelayWhitEnableIsolationFutures.add(maxConsumeDelayFuture);
-        }
-
-        // Wait for all futures to complete
-        CompletableFuture.allOf(maxConsumeDelayWithDisableIsolationFutures.toArray(new CompletableFuture[0])).join();
-        CompletableFuture.allOf(maxConsumeDelayWhitEnableIsolationFutures.toArray(new CompletableFuture[0])).join();
-
-        for (int i = 0; i < loops; i++) {
-            long maxConsumeDelayWithDisableIsolation = maxConsumeDelayWithDisableIsolationFutures.get(i).join();
-            long maxConsumeDelayWhitEnableIsolation = maxConsumeDelayWhitEnableIsolationFutures.get(i).join();
-            log.info("[{}]maxConsumeDelayWithDisableIsolation: {}s, maxConsumeDelayWhitEnableIsolation: {}s", i,
-                    TimeUnit.MILLISECONDS.toSeconds(maxConsumeDelayWithDisableIsolation), TimeUnit.MILLISECONDS.toSeconds(maxConsumeDelayWhitEnableIsolation));
-            assertTrue(TimeUnit.MILLISECONDS.toSeconds(maxConsumeDelayWithDisableIsolation) > sleepTimeMs);
-            if (i == 0) {
-                assertTrue(TimeUnit.MILLISECONDS.toSeconds(maxConsumeDelayWhitEnableIsolation) > sleepTimeMs);
-            } else {
-                assertTrue(TimeUnit.MILLISECONDS.toSeconds(maxConsumeDelayWhitEnableIsolation) < sleepTimeMs);
-            }
-        }
-
-        log.info("-- Exiting {} test --", methodName);
-    }
-
-    private CompletableFuture<Long> startConsumeAndComputeMaxConsumeDelay(String topic, String subscriptionName, Duration consumeSleepTime, boolean enableMessageListenerExecutorIsolation, ExecutorService executorService) throws Exception {
-        int numMessages = 100;
-        final CountDownLatch latch = new CountDownLatch(numMessages);
-        int numPartitions = 100;
-        TopicName nonIsolationTopicName = TopicName.get(topic);
-        admin.topics().createPartitionedTopic(nonIsolationTopicName.toString(), numPartitions);
-
-        AtomicLong maxConsumeDelay = new AtomicLong(-1);
-        ConsumerBuilder<Long> consumerBuilder = pulsarClient.newConsumer(Schema.INT64).topic(nonIsolationTopicName.toString())
-                .subscriptionName(subscriptionName).messageListener((c1, msg) -> {
-                    Assert.assertNotNull(msg, "Message cannot be null");
-                    log.debug("Received message [{}] in the listener", msg.getValue());
-                    c1.acknowledgeAsync(msg);
-                    maxConsumeDelay.set(Math.max(maxConsumeDelay.get(), System.currentTimeMillis() - msg.getValue()));
-                    if (consumeSleepTime.toMillis() > 0) {
-                        Uninterruptibles.sleepUninterruptibly(consumeSleepTime);
-                    }
-                    latch.countDown();
-                });
-
-        ExecutorService executor = Executors.newSingleThreadExecutor(new ExecutorProvider.ExtendedThreadFactory(subscriptionName + "listener-executor-",  true));
-        if (enableMessageListenerExecutorIsolation) {
-            consumerBuilder.messageListenerExecutor((message, runnable) -> executor.execute(runnable));
-        }
-
-        Consumer<Long> consumer = consumerBuilder.subscribe();
-        ProducerBuilder<Long> producerBuilder = pulsarClient.newProducer(Schema.INT64)
-                .topic(nonIsolationTopicName.toString());
-
-        Producer<Long> producer = producerBuilder.create();
-        List<Future<MessageId>> futures = new ArrayList<>();
-
-        // Asynchronously produce messages
-        for (int i = 0; i < numMessages; i++) {
-            Future<MessageId> future = producer.sendAsync(System.currentTimeMillis());
-            futures.add(future);
-        }
-
-        log.info("Waiting for async publish to complete");
-        for (Future<MessageId> future : futures) {
-            future.get();
-        }
-
-        CompletableFuture<Long> maxDelayFuture = new CompletableFuture<>();
-
-        CompletableFuture.runAsync(() -> {
-            try {
-                latch.await(numMessages, TimeUnit.SECONDS);
-            } catch (InterruptedException e) {
-                throw new RuntimeException(e);
-            }
-        }, executorService).whenCompleteAsync((v, ex) -> {
-            maxDelayFuture.complete(maxConsumeDelay.get());
-            try {
-                producer.close();
-                consumer.close();
-                executor.shutdownNow();
-            } catch (PulsarClientException e) {
-                throw new RuntimeException(e);
-            }
-        });
-
-        return maxDelayFuture;
     }
 }
